Fixes Bug 53056: API - erro ao criar cliente quando se indica vendedor com código superior a 2147483647/32767

+++ Sage50c.API.Sample/Sage50c.API.Sample/fAPI.cs
@@ -2066,7 +2066,7 @@ namespace Sage50c.API.Sample {
         private void btnSearchSalesman_Click(object sender, EventArgs e) {
             var SalesmanId = QuickSearchHelper.SalesmanFind();
             if (SalesmanId > 0) {
-                numCustomerSalesmanId.Value = (short)SalesmanId;
+                numCustomerSalesmanId.Value = (long)SalesmanId;
             }
         }
